Introduction documentation

This is an Example project used to test features of the JuliaDocs packages.

section contents

The command contents command inserts a table of contents of all the documentation, as follows:

Introduction documentation

including julia module

including latex in markdown

another chapter

firstt chapter of third part

first section of first chapter of third part

second section of first chapter of third part

but parts are not included ...

including julia module

section include docs

Example.func — Method

func(x)

Returns double the number x plus 1.

Here's an equation:

$\frac{n!}{k!(n - k)!} = \binom{n}{k}$

including latex in markdown

This is text followed by a math environment

\[ f(x) = \int_0^L x dx\]

This is a numbered list where

first entry with text and this math content $x^2$.

second entry with text and this equation

\[ f(x) = \int_0^L x dx\]

third entry. the counter is reseted by the equation

This is a bullet list:

text a

text b with another bulltet list

text c

text d

\[ f(x) = \int_0^L x dx\]

- text e this second level bullet is not interpreted if an equation is inserted before

text f